requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
boto3
pyodbc
python-dotenv
